requirements.txt
beautifulsoup4==4.13.4
certifi==2025.8.3
charset-normalizer==3.4.3
emoji==2.14.1
icalendar==6.3.1
idna==3.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-slugify==8.0.4
pytz==2025.2
requests==2.32.4
six==1.17.0
soupsieve==2.7
text-unidecode==1.3
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
Pillow==11.1.0
